Add the basic implementation AbstractPolymorphism
The code is not full optimized yet. On HackerRank tests this solution
fails for big inputs, lot of set/get operations and big cache size.
I don't know yet what exactly is happening. But probably is fault
of a bad memory management. I didn't deallocate the Nodes on pop.
Probably this is the cause.