Add last solution of Inheritance topic and other from STL
The biggest problem on MagicSpells was not handle the dynamic_cast
to identify the children classes from Spell, was the fucking confusing
implementation of a algorithm to solve the Longest Common Subsequence
between two strings.

As bonus, I added the solution for problem LowerBound on STL. I
want finish all the contests of STL today. Let's me see if I can.